+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/util/Utils.kt
@@ -1,5 +1,6 @@
 package app.cybrid.sdkandroid.util
 
+import android.annotation.SuppressLint
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
@@ -10,22 +11,4 @@ import java.time.format.DateTimeFormatter
 
 fun isSuccessful(code: Int): Boolean {
     return code in 200..299
-}
-
-fun getDateInFormat(date: OffsetDateTime, pattern:String = "MMM dd, YYYY"): String? {
-
-    val formatter = DateTimeFormatter.ofPattern(pattern)
-    return date.format(formatter)
-}
-
-@Composable
-fun getSpannableStyle(
-    text: String,
-    secondaryText: String,
-    style: SpanStyle
-): AnnotatedString {
-    return buildAnnotatedString {
-        append(text)
-        withStyle(style) { append(secondaryText) }
-    }
 }
