+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/util/UtilsComposable.kt
@@ -1,0 +1,29 @@
+package app.cybrid.sdkandroid.util
+
+import android.annotation.SuppressLint
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.withStyle
+import java.time.OffsetDateTime
+import java.time.format.DateTimeFormatter
+
+@SuppressLint("NewApi")
+fun getDateInFormat(date: OffsetDateTime, pattern:String = "MMM dd, YYYY"): String {
+
+    val formatter = DateTimeFormatter.ofPattern(pattern)
+    return date.format(formatter)
+}
+
+@Composable
+fun getSpannableStyle(
+    text: String,
+    secondaryText: String,
+    style: SpanStyle
+): AnnotatedString {
+    return buildAnnotatedString {
+        append(text)
+        withStyle(style) { append(secondaryText) }
+    }
+}
