+++ SDKAndroid/src/test/java/app/cybrid/sdkandroid/util/UtilsTest.kt
@@ -2,6 +2,7 @@ package app.cybrid.sdkandroid.util
 
 import org.junit.Assert
 import org.junit.Test
+import java.time.OffsetDateTime
 
 class UtilsTest {
 
@@ -28,5 +29,23 @@ class UtilsTest {
         Assert.assertTrue(result2)
         Assert.assertFalse(result3)
         Assert.assertFalse(result4)
+    }
+
+    @Test
+    fun getDateInFormatTest() {
+
+        // -- Given
+        val dateOne = OffsetDateTime.parse("2022-08-02T10:55:34.039847-05:00")
+        val patterOne = "MM,dd,YYYY"
+
+        // -- When
+        val dateFormattedOne = getDateInFormat(dateOne)
+        val dateFormattedTwo = getDateInFormat(dateOne, patterOne)
+        val dateFormattedThree = getDateInFormat(dateOne, "")
+
+        // -- Then
+        Assert.assertEquals(dateFormattedOne, "ago. 02, 2022")
+        Assert.assertEquals(dateFormattedTwo, "08,02,2022")
+        Assert.assertEquals(dateFormattedThree, "")
     }
 }
