+++ codecov.yml
@@ -15,6 +15,7 @@ ignore:
   - SDKAndroid/src/main/java/app/cybrid/sdkandroid/components
   - SDKAndroid/src/main/java/app/cybrid/sdkandroid/flow
   - SDKAndroid/src/main/java/app/cybrid/sdkandroid/util/DataAccessStrategy.kt
+  - SDKAndroid/src/main/java/app/cybrid/sdkandroid/util/UtilsCompose.kt
   - SDKAndroid/src/main/java/app/cybrid/sdkandroid/core/Constants.kt
   - SDKAndroid/src/main/java/app/cybrid/sdkandroid/ui/Theme
